Added some things in the Config file, not pushed Added the get used quota for verification Added comment for the simp part and check for no folders error management in upload story introduced main.py has the simps added MainFunctions redesigned so it has removal of pictures, quota check removing of pictures that do not match requirements Added ScheduledSimps.py for extra simp power separately from main

+++ MainFunctions.py
@@ -1,27 +1,38 @@
 import ImageManagement as Im
 import GraphApiUses as Gau
+import Config as Cfg
 import os
 
 
-def upload_picture_on_second_account():
+def upload_picture_on_account_using_api(main_account_used=False):
     response_code = 400
     error_count = 0
     print("\n*******************     New run happened    *******************")
 
-    while response_code != 200 and error_count < 10:
-        photo_folder = "D:\\Exported photos\\Instagram"
-        image_location = Im.get_image_path(photo_folder)
+    while True:
+        if Gau.get_used_quota(False) >= 24:
+            print("The upload quota has been reached, please try later")
+            break
+
+        photo_folder = Cfg.get_parameters()['photo_folder']
         try:
+            image_location = Im.get_image_path(photo_folder)
             while not Im.check_if_valid_ratio(image_location):
+                os.remove(image_location)
                 image_location = Im.get_image_path(photo_folder)
             link = Im.img_bb_image_upload(image_location)
             print("Image is here:", link)
             caption = Im.generate_image_caption(image_location)
-            response = Gau.image_upload(link, caption, main_account=False)
+            response = Gau.image_upload(link, caption, main_account=main_account_used)
             response_code = response.status_code
+            if response_code == 200:
+                print("Photo has been uploaded successfully!")
+                os.remove(image_location)
+                break
         except:
             error_count += 1
             print("Some error happened, error count is ", error_count)
-        if response_code == 200:
-            print("Photo has been uploaded successfully!")
-            os.remove(image_location)
+
+        if error_count > 10:
+            print("Tried 10 times, I give up")
+            break
